refactor(Application): 패키지명 수정
자바 컨벤션에 맞도록 소문자만 사용
