+++ src/main/java/bossmonster/character/Character.java
@@ -1,0 +1,4 @@
+package bossmonster.character;
+
+public class Character {
+}
