+++ src/main/java/bossmonster/resources/Resource.java
@@ -1,4 +1,4 @@
-package bossmonster.VO;
+package bossmonster.resources;
 
 public class Resource {
